re #255 adding some checks for nexus file
