--- docker-compose.yml
@@ -1,64 +1,0 @@
-version: '2'
-
-services:
-
-    producer:
-      image: screamingudder/nexusproducer
-      network_mode: "host"
-      depends_on:
-        - kafka
-        - zookeeper
-
-    kafka:
-      image: wurstmeister/kafka:0.11.0.1
-      depends_on:
-        - zookeeper
-      hostname: kafka
-      restart: unless-stopped
-      ports:
-        - "9092:9092"
-      environment:
-        KAFKA_ADVERTISED_HOST_NAME: localhost
-        KAFKA_ADVERTISED_PORT: 9092
-        KAFKA_ZOOKEEPER_CONNECT: zookeeper:2181
-        KAFKA_MESSAGE_MAX_BYTES: 20000000
-        KAFKA_BROKER_ID: 0
-      volumes:
-        - /var/run/docker.sock:/var/run/docker.sock
-
-    zookeeper:
-      image: zookeeper:3.4
-      restart: unless-stopped
-      ports:
-        - '2181:2181'
-
-    manager:
-      image: sheepkiller/kafka-manager:alpine
-      depends_on:
-        - zookeeper
-        - kafka
-      ports:
-        - "9000:9000"
-      environment:
-        ZK_HOSTS: zookeeper:2181
-
-    # Default logging is user: admin, pass:admin
-    graylog:
-      image: swcc/docker-graylog2
-      ports:
-        - "9001:9000"
-        - "12201:12201"
-
-    file-writer:
-      build: .
-      depends_on:
-        - graylog
-        - kafka
-      network_mode: "host"
-      environment:
-        COMMAND_URI: "//localhost:9092/TEST_writerCommand"
-        STATUS_URI: "//localhost:9092/TEST_writerStatus"
-        KAFKA_BROKER: "localhost"
-        GRAYLOG_ADDRESS: "localhost:12201"
-      volumes:
-        - ./manual-testing/output-files:/output-files
