--- manual-testing/example-json-command.json
@@ -1,37 +1,0 @@
-{
-  "nexus_structure": {
-    "children": [
-      {
-        "type": "group",
-        "name": "for_example_motor_0000",
-        "attributes": {
-          "NX_class": "NXinstrument"
-        },
-        "children": [
-          {
-            "type": "dataset",
-            "name": "some_static_dataset",
-            "values": 42.24,
-            "attributes": {
-              "units": "Kelvin"
-            }
-          },
-          {
-            "type": "stream",
-            "stream": {
-              "topic": "TEST_sampleEnv",
-              "source": "Sample",
-              "module": "f142"
-            }
-          }
-        ]
-      }
-    ]
-  },
-  "file_attributes": {
-    "file_name": "example.nxs"
-  },
-  "cmd": "FileWriter_new",
-  "job_id": "unique-identifier",
-  "broker": "localhost:9092"
-}
