--- manual-testing/output-files/.gitkeep
@@ -1,1 +1,0 @@
-This file retains this otherwise empty directory in git so that it can be used as a file output directory in the docker-compose script.
