--- manual-testing/stop-command.json
@@ -1,4 +1,0 @@
-{
-  "cmd": "FileWriter_stop",
-  "job_id": "unique-identifier"
-}
