--- manual-testing/writer-exit.json
@@ -1,1 +1,0 @@
-{"cmd": "FileWriter_exit"}
