+++ system-tests/helpers/kafkahelpers.py
@@ -1,24 +1,5 @@
-from confluent_kafka import Consumer, Producer
-import uuid
-
-
-def poll_for_valid_message(consumer):
-    """
-    Polls the subscribed topics by the consumer and checks the buffer is not empty or malformed.
-    :param consumer: The consumer object.
-    :return: The message object received from polling.
-    """
-    msg = consumer.poll()
-    assert not msg.error()
-    return msg
-
-
-def create_consumer(offset_reset='smallest'):
-    consumer_config = {'bootstrap.servers': 'localhost:9092',
-                       'default.topic.config': {'auto.offset.reset': offset_reset},
-                       'group.id': uuid.uuid4()}
-    consumer = Consumer(**consumer_config)
-    return consumer
+from confluent_kafka import Producer
+from datetime import datetime
 
 
 def create_producer():
@@ -26,3 +7,16 @@ def create_producer():
                        'message.max.bytes': '20000000'}
     producer = Producer(**producer_config)
     return producer
+
+
+def unix_time_milliseconds(dt):
+    epoch = datetime.utcfromtimestamp(0)
+    return (dt - epoch).total_seconds() * 1000.0
+
+
+def send_writer_command(filepath, producer, topic="TEST_writerCommand"):
+    with open(filepath, "r") as cmd_file:
+        data = cmd_file.read().replace('\n', '')
+        start_time = str(int(unix_time_milliseconds(datetime.utcnow())))
+        data = data.replace('STARTTIME', start_time)
+    producer.produce(topic, data)
